+++ burn-wgpu/src/codegen/dialect/gpu/procedure/mod.rs
@@ -1,9 +1,11 @@
 mod assign;
+mod base;
 mod matmul;
 mod read;
 mod write;
 
 pub(crate) use assign::*;
+pub(crate) use base::*;
 pub(crate) use matmul::*;
 pub(crate) use read::*;
 pub(crate) use write::*;
